requirements.txt
accelerate==1.12.0
altair==6.0.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
diffusers==0.36.0
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.15.0
networkx==3.6.1
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
six==1.17.0
smmap==5.0.2
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.5
zipp==3.23.0
